Refactor group clients for consistent naming and shared functions
