+++ src/app/g/[slug]/(community)/MemberGrid.tsx
@@ -3,6 +3,7 @@
 import React from 'react'
 import type { MemberWithUser } from '@/types'
 import MemberCard from '@/components/MemberCard'
+import { getGridClasses } from '@/lib/group-utils'
 
 interface MemberGridProps {
   members: MemberWithUser[]
@@ -12,27 +13,6 @@ interface MemberGridProps {
   currentUserId?: string
   groupSlug?: string
   gridSize: number
-}
-
-// Helper function to generate dynamic grid classes
-const getGridClasses = (gridSize: number) => {
-  const baseClasses = 'grid gap-4 md:gap-6'
-  
-  // Map gridSize to Tailwind grid-cols classes
-  const gridColsMap: { [key: number]: string } = {
-    1: 'grid-cols-1',
-    2: 'grid-cols-2',
-    3: 'grid-cols-3',
-    4: 'grid-cols-4',
-    5: 'grid-cols-5',
-    6: 'grid-cols-6',
-    7: 'grid-cols-7',
-    8: 'grid-cols-8',
-    9: 'grid-cols-9',
-  }
-  
-  const gridClass = gridColsMap[gridSize] || 'grid-cols-4' // fallback
-  return `${baseClasses} ${gridClass}`
 }
 
 export default function MemberGrid({
